refactor(interpreter): Split out `Element` and `TagName`
